Optimisation et finition de l'API

+++ app/Http/Controllers/CommentController.php
@@ -2,19 +2,23 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\NotFoundException;
+use App\Exceptions\ValidatorException;
 use App\Http\Resources\BaseResource;
 use App\Http\Resources\UserResource;
+use App\Models\Post;
 use Illuminate\Http\Request;
 use App\Models\Comment;
 use App\Models\User;
 use App\Http\Resources\CommentCollection;
 use App\Http\Resources\CommentResource;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class CommentController extends Controller
 {
     public function index() {
-        $comments = Comment::where('post_type', "comment")->get();
+        $comments = Comment::all();
         $collection = new CommentCollection($comments);
 
         return $collection
@@ -23,22 +27,39 @@ class CommentController extends Controller
             ->setMessage("Comment list loaded successfully");
     }
 
-    public function store($id, Request $request) {
+    public function show($id) {
+        $comment = Comment::where("id", $id)
+//            ->with("author")
+            ->where("post_type", "comment")
+            ->firstOrFail();
+        $resource = new CommentResource($comment);
+
+        return $resource
+            ->success()
+            ->setCode(200)
+            ->setMessage("Comment retrieved successfully");
+    }
+
+    public function store(int $id, Request $request) {
+        $post = Post::find($id);
+
+        if (!$post) {
+            throw new NotFoundException("Not found", json_encode(["not_found"=> "The post you're trying to add a comment does not exist"]));
+        }
+
+        $this->authorize("create", [Comment::class, $post]);
+
         $validator = Validator::make($request->all(), [
             "comment_content"=> "required|string",
-            "comment_author"=> "required|integer",
         ]);
 
         if ($validator->fails()) {
-            return BaseResource::error()
-                ->setCode(400)
-                ->setMessage("Data validation failed")
-                ->setErrors($validator->errors());
+            throw new ValidatorException("Data validation failed", json_encode($validator->errors()));
         }
 
         $data = array(
             "post_content"=> $request->get("comment_content"),
-            "post_author"=> $request->get("comment_author"),
+            "post_author"=> Auth::id(),
             "post_type"=> "comment",
             "post_parent"=> $id
         );
@@ -52,25 +73,12 @@ class CommentController extends Controller
             ->setMessage("Comment created successfully");
     }
 
-    public function show($id) {
-        $comment = Comment::where("id", $id)
-            ->where("post_type", "comment")
-            ->firstOrFail();
-        $resource = new CommentResource($comment);
-
-        return $resource
-            ->success()
-            ->setCode(200)
-            ->setMessage("Comment retrieved successfully");
-    }
-
     public function destroy($id) {
         $comment = Comment::find($id);
 
-        if (!$comment) return BaseResource::error()
-            ->setCode(404)
-            ->setMessage("Comment not found")
-            ->setErrors(json_encode(["not_found"=> "The comment you're trying to delete does not exist"]));
+        if (!$comment) throw new NotFoundException("Not found", json_encode(["not_found"=> "The comment you're trying to delete does not exist"]));
+
+        $this->authorize("delete", $comment);
 
         $comment->delete();
 
@@ -79,18 +87,6 @@ class CommentController extends Controller
         return BaseResource::error()
             ->success()
             ->setCode(200)
-            ->setMessage("Comment (" . $id . ") `" . $this->truncateString($comment->post_content, 20) . "` posted by `" . $author->user_display_name . "` deleted successfully");
-    }
-
-    function truncateString(string $input, int $limit): string
-    {
-        // Vérifie si la chaîne est plus longue que la limite
-        if (strlen($input) > $limit) {
-            // Coupe la chaîne à la limite spécifiée et ajoute "..."
-            return substr($input, 0, $limit) . '...';
-        }
-
-        // Retourne la chaîne si elle est déjà inférieure ou égale à la limite
-        return $input;
+            ->setMessage("Comment (" . $id . ") `" . truncateString($comment->post_content, 20) . "` posted by `" . $author->user_display_name . "` deleted successfully");
     }
 }
