+++ app/Http/Controllers/PostController.php
@@ -2,18 +2,21 @@
 
 namespace App\Http\Controllers;
 
+use App\Exceptions\NotFoundException;
+use App\Exceptions\ValidatorException;
 use App\Http\Resources\BaseResource;
 use Illuminate\Http\Request;
 use App\Models\Post;
 use App\Http\Resources\PostCollection;
 use App\Http\Resources\PostResource;
 use App\Models\Topic;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class PostController extends Controller
 {
     public function index($id = null) {
-        $posts = Post::where("post_type", "post")->get();
+        $posts = Post::with("author")->where("post_type", "post")->get();
         $collection = new PostCollection($posts);
 
         return $collection
@@ -22,7 +25,10 @@ class PostController extends Controller
             ->setMessage("Post list loaded successfully");
     }
     public function show($id) {
-        $post = Post::findOrFail($id);
+        $post = Post::where("post_type", "post")->find($id);
+
+        if (!$post) throw new NotFoundException("Not found", json_encode(["not_found"=> "The post you're trying to access does not exist"]));
+
         $resource = new PostResource($post);
 
         return $resource
@@ -31,23 +37,25 @@ class PostController extends Controller
             ->setMessage("Post retrieved successfully");
     }
 
-    public function store(Request $request) {
+    public function store(Request $request, int $id) {
+        $topic = Topic::find($id);
+
+        if (!$topic) throw new NotFoundException("Not found", json_encode(["not_found"=> "The topic you're trying to access does not exist"]));
+
+        //  Check user permissions to create post
+        $this->authorize("create", [Post::class, $topic]);
+
 
         $validator = Validator::make($request->all(), [
             "post_title"=> "required|string",
             "post_description"=> "required|string",
             "post_content"=> "required|string",
             "post_coordinates_lat"=> "decimal:1,10|nullable",
-            "post_coordinates_long"=> "decimal:1,10|nullable",
-            "post_author"=> "required|integer",
-            "post_parent"=> "integer|nullable"
+            "post_coordinates_long"=> "decimal:1,10|nullable"
         ]);
 
         if($validator->fails()) {
-            return BaseResource::error()
-                ->setCode(400)
-                ->setMessage("Data validation failed")
-                ->setErrors($validator->errors());
+            throw new ValidatorException("Data validation failed", $validator->errors());
         }
 
         $data = array(
@@ -57,8 +65,8 @@ class PostController extends Controller
             "post_content"=> $request->get("post_content"),
             "post_coordinates"=> serialize(array("lat"=> $request->get("post_coordinates_lat"), "long"=> $request->get("post_coordinates_long"))),
             "post_type"=> "post",
-            "post_author"=> $request->get("post_author"),
-            "post_parent"=> $request->get("post_parent")
+            "post_author"=> Auth::id(),
+            "post_parent"=> $topic->id
         );
 
         $post = Post::create($data);
@@ -70,7 +78,10 @@ class PostController extends Controller
             ->setMessage("Post created successfully");
     }
 
-    public function update(Request $request, $id) {
+    public function update(Request $request, int $id) {
+        $old_post = Post::findOrFail($id);
+
+        $this->authorize("update", [$old_post]);
 
         $validator = Validator::make($request->all(), [
             "post_title"=> "string|nullable",
@@ -79,14 +90,10 @@ class PostController extends Controller
         ]);
 
         if ($validator->fails()) {
-            return BaseResource::error()
-                ->setCode(400)
-                ->setMessage("Data validation failed")
-                ->setErrors($validator->errors());
+            throw new ValidatorException("Data validation failed", json_encode($validator->errors()));
         }
 
         $post = Post::findOrFail($id);
-        $old_post = Post::findOrFail($id);
 
         $fieldsToUpdate = [
             'post_title' => 'post_title',
@@ -126,10 +133,9 @@ class PostController extends Controller
 
         $post = Post::find($id);
 
-        if (!$post) return BaseResource::error()
-            ->setCode(404)
-            ->setMessage("Post not found")
-            ->setErrors(json_encode(["not_found"=> "The post you're trying to delete does not exist"]));
+        if (!$post) throw new NotFoundException("Not Found", json_encode(["not_found"=> "The post you're trying to delete does not exist"]));
+
+        $this->authorize("delete", [$post]);
 
         $post->delete();
 
@@ -139,16 +145,14 @@ class PostController extends Controller
             ->setMessage("Post (" . $id . ") `" . $post->post_title . "` deleted successfully");
     }
 
-    public function getPostsByTopic($id): PostCollection
-    {
-        $topic = Topic::findOrFail($id);
+    public function getPostsMap($location = null) {
+        // TODO : Logic to get all close point by user location
+        $posts = Post::with("attachments")->get();
+        $collection = new PostCollection($posts);
 
-        $posts = $topic->posts()->where('post_type', 'post')->get();
-
-
-        return (new PostCollection($posts))
+        return $collection
             ->success()
             ->setCode(200)
-            ->setMessage("Posts for this topic retrieved successfully");
+            ->setMessage("Post list loaded successfully");
     }
 }
